refactor(storage): split lib.rs into store/ directory module
Drains the 4509-line src/crates/primer-storage/src/lib.rs into a
directory-module split that mirrors the dialogue_manager/ pattern.
lib.rs is now a 24-line shim; every production file under store/ is
under 500 code lines (largest: session_save.rs at 399).

  store/mod.rs          81 LOC  struct + open_for_locale + locale
  store/session.rs     120 LOC  impl SessionStore — one-line delegations
  store/session_save.rs 399 LOC pub(super) *_inner save methods
  store/session_load.rs 214 LOC pub(super) *_inner load methods
  store/session_search.rs 212 LOC BM25 + hybrid + knn helper
  store/learner.rs     314 LOC  impl LearnerStore (save + load)
  store/embeddings.rs  152 LOC  embedding-model registry + vec ops
  store/conv.rs         76 LOC  pure type conversions + tests
  store/fts.rs          65 LOC  pure FTS5 sanitisation + tests

The SessionStore trait impl stays a single block (in session.rs) and
each method one-line-delegates to a pub(super) async fn *_inner
inherent method on SqliteSessionStore living in the appropriate axis
file. This preserves &self access to self.locale and self.conn,
preserves the single conn.lock() per public entry point invariant, and
keeps the #[deprecated] open() shim path intact.

Pure helpers (parse_rfc3339, i64_to_u8/u32/u64, sanitize_fts_phrase,
vec_to_blob, blob_to_vec, cosine, validate_storage_embedder,
upsert_storage_embedding_model) lift cleanly into reusable modules
with co-located tests, mirroring the apply.rs pattern. cosine /
vec_to_blob / blob_to_vec are tagged with a TODO — they're duplicated
verbatim in primer-knowledge; lifting to primer-core::vector is a
deferred follow-up.

Tests moved from two #[cfg(test)] mod blocks at the bottom of lib.rs
into store/tests/{session_tests,learner_tests}.rs, declared via
`#[cfg(test)] mod tests { mod ...; }` from store/mod.rs. Further
per-axis test split (schema/migrations/save/retrieval/hybrid) is a
deferred follow-up. catalog.rs (361 LOC) and schema.rs (959 LOC) are
unchanged and out of scope.

Pure code motion. Public API unchanged. 147 tests pass (124 pre-existing
+ 23 new co-located helper tests); workspace still 584 passing.
